+++ concurrency/src/test/java/co/kr/woojjam/concurrency/entity/TestCoupon.java
@@ -1,4 +1,4 @@
-package co.kr.woojjam.concurrency.coupon;
+package co.kr.woojjam.concurrency.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -22,26 +22,24 @@ public class TestCoupon {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	private String code; // 쿠폰 코드
-	private int discountAmount; // 할인 금액
 	private int stock; // 남은 쿠폰 개수
 
-	@Version
-	private Integer version;
+	// @Version
+	// private Integer version;
 
 	@Builder
-	public TestCoupon(final Long id, final String code, final int discountAmount, final int stock,
+	public TestCoupon(final Long id, final String code, final int stock,
 		final Integer version) {
 		this.id = id;
 		this.code = code;
-		this.discountAmount = discountAmount;
 		this.stock = stock;
-		this.version = version;
+		// this.version = version;
 	}
 
 	public void useCoupon() {
-		if (this.stock <= 0) {
+		if (this.stock > 5) {
 			throw new IllegalStateException("쿠폰 재고가 부족합니다.");
 		}
-		this.stock -= 1;
+		this.stock += 1;
 	}
 }
