+++ concurrency/src/test/java/co/kr/woojjam/concurrency/repository/TestCouponRepository.java
@@ -1,8 +1,8 @@
-package co.kr.woojjam.concurrency.coupon;
+package co.kr.woojjam.concurrency.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
+import co.kr.woojjam.concurrency.entity.TestCoupon;
+
 public interface TestCouponRepository extends JpaRepository<TestCoupon, Long> {
 }
